requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy>=2.0.21
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation & Serialization
pydantic>=2.4.2
pydantic-settings>=2.0.3

# Caching
redis==5.0.1

# HTTP Client
httpx==0.25.2

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2